fixed pr comments pt2
